requirements.txt
flask>=3.0
gunicorn>=21.2
python-pptx>=0.6.23
pydantic>=2.0
requests>=2.31
google-genai>=1.51.0
Pillow>=10.0
python-dotenv>=1.0
